% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_trans.R
\name{trans_affine}
\alias{trans_affine}
\title{Apply an affine transformation to a set of points}
\usage{
trans_affine(data, center = c(0, 0), scale = diag(1, 1), r_theta_rad = 0)
}
\arguments{
\item{data}{data frame containing (at least) columns \code{x} and \code{y}}

\item{center}{length-2 vector containing the (x, y) offset}

\item{scale}{2x2 matrix, possibly diagonal.}

\item{r_theta_rad}{rot}
}
\value{
data frame, same number of rows as \code{data} with modified \code{x} and \code{y}.
}
\description{
The points will be
\enumerate{
\item Scaled by the matrix \code{scale},
\item Rotated through an angle \code{r_theta_rad} about the origin,
\item Translated by \code{center}.
}
}
\details{
\code{scale} doesn't have to be a diagonal matrix, in which case it
is just applied as a linear transformation.
}
\examples{
cir = unit_circle(ppoints(20)) \%>\% select(-t)
plot(cir)
ell = trans_affine(cir, scale  = diag(c(2, 1)), r_theta_rad = pi/4, center = c(1, 1))
plot(ell, col = 'red', add = TRUE)
}
